+++ packages/storybook/builders.json
@@ -1,14 +1,25 @@
 {
-  "$schema": "@angular-devkit/architect/src/builders-schema.json",
   "builders": {
     "storybook": {
-      "implementation": "./src/builders/storybook/storybook.impl",
-      "schema": "./src/builders/storybook/schema.json",
+      "implementation": "./src/executors/storybook/compat",
+      "schema": "./src/executors/storybook/schema.json",
       "description": "Serve Storybook"
     },
     "build": {
-      "implementation": "./src/builders/build-storybook/build-storybook.impl",
-      "schema": "./src/builders/build-storybook/schema.json",
+      "implementation": "./src/executors/build-storybook/compat",
+      "schema": "./src/executors/build-storybook/schema.json",
+      "description": "Build Storybook"
+    }
+  },
+  "executors": {
+    "storybook": {
+      "implementation": "./src/executors/storybook/storybook.impl",
+      "schema": "./src/executors/storybook/schema.json",
+      "description": "Serve Storybook"
+    },
+    "build": {
+      "implementation": "./src/executors/build-storybook/build-storybook.impl",
+      "schema": "./src/executors/build-storybook/schema.json",
       "description": "Build Storybook"
     }
   }
