--- packages/storybook/src/builders/storybook/storybook.impl.spec.ts
@@ -1,5 +1,0 @@
-describe('storybook builer', () => {
-  it('should have a test', () => {
-    expect(true).toBeTruthy();
-  });
-});
