+++ packages/storybook/src/executors/build-storybook/compat.ts
@@ -1,0 +1,4 @@
+import { convertNxExecutor } from '@nrwl/devkit';
+import buildStorybookExecutor from './build-storybook.impl';
+
+export default convertNxExecutor(buildStorybookExecutor);
