+++ packages/storybook/src/executors/storybook/compat.ts
@@ -1,0 +1,4 @@
+import { convertNxExecutor } from '@nrwl/devkit';
+import storybookExecutor from './storybook.impl';
+
+export default convertNxExecutor(storybookExecutor);
