+++ packages/storybook/src/executors/storybook/schema.json
@@ -1,5 +1,6 @@
 {
   "title": "Storybook Dev Builder",
+  "cli": "nx",
   "description": "Serve up storybook in development mode",
   "type": "object",
   "properties": {
