+++ packages/storybook/src/utils/utilities.ts
@@ -1,12 +1,6 @@
 import { Tree } from '@nrwl/devkit';
 
-import { get } from 'http';
 import { CompilerOptions } from 'typescript';
-
-export interface NodePackage {
-  name: string;
-  version: string;
-}
 
 export const Constants = {
   addonDependencies: ['@storybook/addons'],
@@ -51,41 +45,6 @@ export function safeFileDelete(tree: Tree, path: string): boolean {
     return true;
   } else {
     return false;
-  }
-}
-
-/**
- * Attempt to retrieve the latest package version from NPM
- * Return an optional "latest" version in case of error
- * @param packageName
- */
-export function getLatestNodeVersion(
-  packageName: string
-): Promise<NodePackage> {
-  const DEFAULT_VERSION = 'latest';
-
-  return new Promise((resolve) => {
-    return get(`http://registry.npmjs.org/${packageName}`, (res) => {
-      let rawData = '';
-      res.on('data', (chunk) => (rawData += chunk));
-      res.on('end', () => {
-        try {
-          const response = JSON.parse(rawData);
-          const version = (response && response['dist-tags']) || {};
-
-          resolve(buildPackage(packageName, version.latest));
-        } catch (e) {
-          resolve(buildPackage(packageName));
-        }
-      });
-    }).on('error', () => resolve(buildPackage(packageName)));
-  });
-
-  function buildPackage(
-    name: string,
-    version: string = DEFAULT_VERSION
-  ): NodePackage {
-    return { name, version };
   }
 }
